+++ sphinx/source/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 * Converted the :ref:`htmlTemplates` from XHTML 1.0 to HTML5
 * Added the :ref:`t_footer` template (for formatting the ``<footer>`` element
   of a page)
+* Added the :ref:`assemble` ASM directive
 
 4.5 (2015-07-04)
 ----------------